requirements.txt
pathlib
Pillow
imagehash
pyinstaller